create `DarkModeSelector` component
